[WAYANG-#8] move old code to folder out of the current pywy
