+++ python/old_code/old_graph/node.py
