+++ python/old_code/old_graph/visitant.py
