+++ python/old_code/tests/full_spark_test.py
@@ -14,10 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import unittest
-from pywy.orchestrator.plan import Descriptor
-from pywy.orchestrator.dataquanta import DataQuantaBuilder
-
+#
 
 def test_most_basic(self):
     descriptor = Descriptor()
